+++ doc/version_history.rst
@@ -6,6 +6,17 @@
 Version History
 ###############
 
+v0.5.0
+======
+
+* Prepare for salobj 7.
+
+Requires:
+
+* ts_salobj 7
+* ts_idl 3.1
+* IDL file for GenericCamera from ts_xml 11
+
 v0.4.0
 ======